+++ apps/docs/src/components/Inputs/textfield/MainView.tsx
@@ -1,4 +1,5 @@
-import { HelpText, TextField, VBox } from "@elf-framework/ui";
+import React from "react";
+import { HelpText, TextField, VBox } from "@react-elf/ui";
 
 export function MainView() {
   return (
@@ -8,6 +9,10 @@ export function MainView() {
         backgroundColor: "var(--color-gray-0)",
         padding: [60, 300],
         position: "relative",
+        fontSize: 13,
+        justifyContent: "center",
+        alignItems: "center",
+        height: 300,
       }}
     >
       <div