+++ packages/@react-elf/input-editor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @react-elf/input-editor
 
+## 0.0.86
+
+### Patch Changes
+
+- update TextField component
+
+- Updated dependencies []:
+  - @react-elf-types/input-editor@0.0.86
+
 ## 0.0.85
 
 ### Patch Changes