+++ packages/@react-elf/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-elf/button
 
+## 0.0.86
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf/input-editor@0.0.86
+
 ## 0.0.85
 
 ### Patch Changes